set line height & spacing for buffer
Using `default-text-properties'

As far as I could determine, there was no way to set line height for unmatched
elements without using default text properties.  The buffer local variable for
`line-spacing` makes it impossible to set a lower value for source blocks.  This
is mainly what drove the implementation choice.

Font lock will continue to manage the line height and spacing even if org modern
is turned off.  Cleaning this up is possible but was judged to be harmless to
leave behind.  There are certain cases where "cleanup" and state can become
inconsistent anyway.