New box-face methodology, associated faces, and tooling
We now measure the appearance of 6 types of entity types in default
text and all org headings directly in a temporary buffer in the
current window.  Those are saved into new org-modern--box-label-type-N
faces, to which properties org-modern-raise and org-modern-fixed are
applied, the latter two of which can be used for 'display properties.
Face remappings for these faces are applied on text-scale changes.
Centered box labeling can now also be applied to labels outside
headlines.