Do not complain about :height=1.0 on fixed height faces
The display raise property is relative to the height of the text being
raised.  For fixed height labels, it's important for that text to be
uniform in height and not adapt to the text around. In practice this
means omitting any :height face attribute on that text.  We check for
that, but :height=1.0 is essentially the same as it not appearing.
Unfortunately for some reason the fixed-pitch face specifies
:height=1.0.